Swift: move `TargetFile` to a separate lib
