+++ swift/extractor/infra/file/BUILD.bazel
@@ -1,13 +1,9 @@
 load("//swift:rules.bzl", "swift_cc_library")
 
 swift_cc_library(
-    name = "infra",
+    name = "file",
     srcs = glob(["*.cpp"]),
     hdrs = glob(["*.h"]),
     visibility = ["//swift:__subpackages__"],
-    deps = [
-        "//swift/extractor/infra/file",
-        "//swift/extractor/trap",
-        "//swift/third_party/swift-llvm-support",
-    ],
+    deps = ["//swift/third_party/swift-llvm-support"],
 )
