+++ swift/extractor/infra/file/TargetFile.cpp
@@ -1,4 +1,4 @@
-#include "swift/extractor/infra/TargetFile.h"
+#include "swift/extractor/infra/file/TargetFile.h"
 
 #include <iostream>
 #include <cstdio>
